TEMP: Fix camera screen in landscape orientation
Currently photos taken in Adroit in portrait mode are displayed correctly in Adroit as portrait photos but in the Photos app they are rotated 90 degrees. This is a known issue with React Native/react-native-camera which only affects certain iOS and Samsung devices. We'll potentially revisit this and enable taking portrait photos at a later stage.